create a check to make sure a contact belongs to a user before updating or deleting a contact

+++ controllers/contactController.js
@@ -64,6 +64,11 @@ const updateContact = asyncHandler(async (req, res) => {
     throw new Error('Contact not found')
   }
 
+  if(contact.user_id.toString() !== req.user.id) {
+    res.status(403)
+    throw new Error('User does not have permission to update other users contacts')
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body, {new: true})
 
   res.status(200).json(updatedContact)
@@ -78,6 +83,11 @@ const deleteContact = asyncHandler(async (req, res) => {
   if(!contact) {
     res.status(404)
     throw new Error('Contact not found')
+  }
+
+  if(contact.user_id.toString() !== req.user.id) {
+    res.status(403)
+    throw new Error('User does not have permission to update other users contacts')
   }
 
   await Contact.deleteOne(contact)
